add conda and usegalaxy badge
Vardict is now available in Bioconda and in the [European Galaxy server](https://usegalaxy.eu). The Europan Galaxy server is part of the `German Network of Bioinformatics Infrastructure` (de.NBI), ELIXIR, and the European Science Cloud (EOSC-life). We hope that this badge will increase the `vardict` userbase in the community by pointing user to a combination of a graphical user-interface and compute resources.